+++ Samples/2.0/Tutorials/Tutorial06_Multithreading/Tutorial06_Multithreading.cpp
@@ -147,7 +147,7 @@ unsigned long renderThreadApp( Ogre::ThreadHandle *threadHandle )
             Ogre::Threads::Sleep( 120 );
 
         Ogre::uint64 endTime = timer.getMicroseconds();
-        timeSinceLast = ( endTime - startTime ) / 1000000.0;
+        timeSinceLast = double( endTime - startTime ) / 1000000.0;
         timeSinceLast = std::min( 1.0, timeSinceLast );  // Prevent from going haywire.
         startTime = endTime;
     }
